requirements.txt
asgiref==3.8.1
Django==5.1.3
django-restframework==0.0.1
sqlparse==0.5.2
